requirements.txt
flask
moviepy
gTTS
requests
Pillow
